pyproject.toml
[project]
name = "Ygg"
version = "0.1.0"
license = "MIT"
license-files = ["LICENSE"]
description = "Implements dynamic data contracts based on the Open Data Contract Standards."
readme = "README.md"
authors = [
  {name = "Thiago Dias", email = "thiago@yggdrasil.codes"}
]
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.39",
    "boto3-stubs~=1.42.39",
    "dotenv>=0.9.9",
    "duckdb>=1.4.3",
    "glom>=25.12.0",
    "ipykernel>=7.1.0",
    "jinja2>=3.1.6",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml==6.0.3",
    "ruff>=0.14.10",
    "snowflake-connector-python>=4.2.0",
    "sqlmodel>=0.0.33",
    "streamlit>=1.53.1",
    "structlog>=25.5.0",
    "tabulate>=0.9.0",
]

[tool.setuptools.package-data]
"ygg" = ["*.yaml", "assets/**/*.yaml", "*.json", "assets/**/*.json"]
